backend/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2.3
django-colorfield==0.14.0
djangorestframework==3.16.0
pillow==11.2.1
psycopg2==2.9.10
sqlparse==0.5.3
